dpdk: fix overflow in mtu arithmetic
When the driver's max_rx_pktlen is >= 65536, max_supported_frame_size
overflows and queue creation fails.

Type: fix

+++ src/plugins/dpdk/device/common.c
@@ -65,7 +65,7 @@ dpdk_device_setup (dpdk_device_t * xd)
   struct rte_eth_dev_info dev_info;
   struct rte_eth_conf conf = {};
   u64 rxo, txo;
-  u16 max_frame_size;
+  u32 max_frame_size;
   int rv;
   int j;
